backend/requirements.txt
google-api-python-client>=2.131.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.1
phonenumbers>=8.13.37
schedule>=1.2.1
requests>=2.32.3
pydantic>=2.7.1
